Fixup min/max functions in ordered.
